add build tags to all integration tests
