--- tests/integration/aliases/aliases_test.go
@@ -1,104 +1,0 @@
-// Copyright 2016-2018, Pulumi Corporation.  All rights reserved.
-
-package ints
-
-import (
-	"path/filepath"
-	"testing"
-
-	"github.com/pulumi/pulumi/pkg/v2/testing/integration"
-)
-
-var dirs = []string{
-	"rename",
-	"adopt_into_component",
-	"rename_component_and_child",
-	"retype_component",
-	"rename_component",
-}
-
-// TestNodejsAliases tests a case where a resource's name changes but it provides an `alias`
-// pointing to the old URN to ensure the resource is preserved across the update.
-func TestNodejsAliases(t *testing.T) {
-	for _, dir := range dirs {
-		d := filepath.Join("nodejs", dir)
-		t.Run(d, func(t *testing.T) {
-			integration.ProgramTest(t, &integration.ProgramTestOptions{
-				Dir:          filepath.Join(d, "step1"),
-				Dependencies: []string{"@pulumi/pulumi"},
-				Quick:        true,
-				EditDirs: []integration.EditDir{
-					{
-						Dir:             filepath.Join(d, "step2"),
-						Additive:        true,
-						ExpectNoChanges: true,
-					},
-				},
-			})
-		})
-	}
-}
-
-func TestPythonAliases(t *testing.T) {
-	for _, dir := range dirs {
-		d := filepath.Join("python", dir)
-		t.Run(d, func(t *testing.T) {
-			integration.ProgramTest(t, &integration.ProgramTestOptions{
-				Dir: filepath.Join(d, "step1"),
-				Dependencies: []string{
-					filepath.Join("..", "..", "..", "sdk", "python", "env", "src"),
-				},
-				Quick: true,
-				EditDirs: []integration.EditDir{
-					{
-						Dir:             filepath.Join(d, "step2"),
-						Additive:        true,
-						ExpectNoChanges: true,
-					},
-				},
-			})
-		})
-	}
-}
-
-func TestDotNetAliases(t *testing.T) {
-	for _, dir := range dirs {
-		d := filepath.Join("dotnet", dir)
-		t.Run(d, func(t *testing.T) {
-			integration.ProgramTest(t, &integration.ProgramTestOptions{
-				Dir:          filepath.Join(d, "step1"),
-				Dependencies: []string{"Pulumi"},
-				Quick:        true,
-				EditDirs: []integration.EditDir{
-					{
-						Dir:             filepath.Join(d, "step2"),
-						Additive:        true,
-						ExpectNoChanges: true,
-					},
-				},
-			})
-		})
-	}
-}
-
-func TestGoAliases(t *testing.T) {
-	for _, dir := range dirs {
-		d := filepath.Join("go", dir)
-		t.Run(d, func(t *testing.T) {
-			integration.ProgramTest(t, &integration.ProgramTestOptions{
-				Dir: filepath.Join(d, "step1"),
-				Dependencies: []string{
-					"github.com/pulumi/pulumi/sdk/v2",
-				},
-				Quick: true,
-				EditDirs: []integration.EditDir{
-					{
-						Dir:             filepath.Join(d, "step2"),
-						ExpectNoChanges: true,
-						Additive:        true,
-					},
-				},
-			})
-		})
-	}
-}
