+++ tests/integration/custom_timeouts/custom_timeouts_test.go
@@ -1,3 +1,4 @@
+// +build python all
 package ints
 
 import (
