+++ tests/integration/duplicate_urns/duplicate_urns_test.go
@@ -1,4 +1,5 @@
 // Copyright 2016-2018, Pulumi Corporation.
+// +build nodejs all
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
