+++ tests/integration/read/import_acquire/import_acquire_test.go
@@ -1,4 +1,5 @@
 // Copyright 2016-2018, Pulumi Corporation.  All rights reserved.
+// +build nodejs all
 
 package ints
