+++ tests/integration/steps/steps_test.go
@@ -1,4 +1,5 @@
 // Copyright 2016-2018, Pulumi Corporation.  All rights reserved.
+// +build nodejs all
 
 package ints
