+++ tests/integration/transformations/transformations_nodejs_test.go
@@ -1,0 +1,29 @@
+// Copyright 2016-2020, Pulumi Corporation.  All rights reserved.
+// +build nodejs all
+
+package ints
+
+import (
+	"path/filepath"
+	"testing"
+
+	"github.com/pulumi/pulumi/pkg/v2/testing/integration"
+)
+
+var dirs = []string{
+	"simple",
+}
+
+func TestNodejsTransformations(t *testing.T) {
+	for _, dir := range dirs {
+		d := filepath.Join("nodejs", dir)
+		t.Run(d, func(t *testing.T) {
+			integration.ProgramTest(t, &integration.ProgramTestOptions{
+				Dir:                    d,
+				Dependencies:           []string{"@pulumi/pulumi"},
+				Quick:                  true,
+				ExtraRuntimeValidation: Validator("nodejs"),
+			})
+		})
+	}
+}
