+++ tests/integration/types/types_test.go
@@ -1,4 +1,5 @@
 // Copyright 2016-2020, Pulumi Corporation.  All rights reserved.
+// +build python all
 
 package ints
